Split DatasetCollectionConfig to own class.
